==> project/notebooks/0_index.ipynb <==
# CSC 466 Final Analytical Project (Winter 2025)
- Lucas Summers (lsumme01@calpoly.edu)
- Xiuyuan Qiu (xiqiu@calpoly.edu)
- Braeden Alonge (balonge@calpoly.edu)
- Nathan Lim (nlim10@calpoly.edu)

# Board Game Geek Dataset Analysis

This project analyzes a comprehensive dataset of board games to understand what factors contribute to high ratings. The analysis uses principal component analysis, random forest/linear regression, collaborative filtering, and various clustering techniques to identify patterns and insights.

## Notebook Structure

1. [Data Preparation](./1_data_preparation.ipynb): Loading and preprocessing the board game dataset.
2. [PCA Analysis](./2_pca_analysis.ipynb): Reducing dimensionality of binary board game features.
3. [High Ratings Clustering](./3_high_ratings.ipynb): Analyzing clusters within highly-rated board games (rating >= 7.0).
4. [Low Ratings Clustering](./4_low_ratings.ipynb): Analyzing clusters within low-rated board games (rating < 6.0).
5. [High vs. Low Clustering Comparison](./5_high_low_comparison.ipynb): Analyzing the differences between high-rated and low-rated board games.
6. [Collaborative Filtering](./6_collaborative_filtering.ipynb): Exploration of a collaborative filtering model based on the user_ratings data to recommend games to users and suggest similar games.
7. [Predicting Average Rating](./7_predict_average_rating.ipynb): Train a Linear Regression and Random Forest models to predict average rating for a game.

## How to Run

Execute the notebooks in order to reproduce the analysis. **Some notebooks load data saved by previous notebooks.**

- Dataset can be found [here](https://www.kaggle.com/datasets/threnjen/board-games-database-from-boardgamegeek/) and all CSV files must be in the `data/` directory at the root of the project directory
- Intermediate dataframes and data shared between notebooks are stored in the `frames/` directory
- All plots generated in notebooks are stored in the `plots/` directory
